# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "NimeStream API documentation",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/app/banner": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner_app"
                ],
                "summary": "Get all banner_app",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Type of banner (e.g., 'anime', 'movie', kdrama)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner_app"
                ],
                "summary": "Create a new banner_app",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_banner_app_request.CreateBannerApp"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/app/banner/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner_app"
                ],
                "summary": "Get a banner_app by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BannerApp ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner_app"
                ],
                "summary": "Update a banner_app",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BannerApp ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_banner_app_request.UpdateBannerApp"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner_app"
                ],
                "summary": "Delete a banner_app",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BannerApp ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/movie/details": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Get all movie_details",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term like title or studio",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "anime",
                        "description": "Type of movie",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved data",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithPaginate-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieDetails"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Create a new movie_details",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_details_request.CreateMovieDetails"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Data created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieDetails"
                        }
                    }
                }
            }
        },
...
        "/movie/details/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Get a movie_details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MovieDetails ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieDetails"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Update a movie_details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MovieDetails ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_details_request.UpdateMovieDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieDetails"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Delete a movie_details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MovieDetails ID (uint)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_banner_app_request.CreateBannerApp": {
            "type": "object",
            "properties": {
                "banner_type": {
                    "description": "e.g., \"movie\", \"drakor\", \"anime\"",
                    "type": "string"
                },
                "detail_url": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_banner_app_request.UpdateBannerApp": {
            "type": "object",
            "properties": {
                "banner_type": {
                    "type": "string"
                },
                "detail_url": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_details_request.CreateMovieDetails": {
            "type": "object",
            "required": [
                "genres",
                "movie_id",
                "movie_type",
                "thumbnail_url",
                "title"
            ],
            "properties": {
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "demon",
                        "fantasy",
                        "supernatural"
                    ]
                },
                "movie_id": {
                    "type": "string",
                    "example": "wu-nao-monu"
                },
                "movie_type": {
                    "type": "string",
                    "example": "anime"
                },
                "producer": {
                    "type": "string",
                    "example": "Agate"
                },
                "rating": {
                    "type": "string",
                    "example": "5.9"
                },
                "release_date": {
                    "type": "string",
                    "example": "2023-05-01"
                },
                "status": {
                    "type": "string",
                    "example": "complete"
                },
                "studio": {
                    "type": "string",
                    "example": "Agate"
                },
                "synopsis": {
                    "type": "string",
                    "example": "Agate is a demon girl cursed with an eternal life. To forget the past, she throws half of her head into a deep valley and runs away. Unexpectedly, the tears flood the valley floor and form into a lake, triggering a flood in hell. Aloys, the Prince of Ghost comes to the Human World to find out the truth and finally finds the demon girl, starting a story of life and love."
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://s4.anilist.co/file/anilistcdn/media/anime/cover/medium/bx141914-P1xQHMXN7q6z.png"
                },
                "title": {
                    "type": "string",
                    "example": "Wu Nao Monü"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_details_request.UpdateMovieDetails": {
            "type": "object",
            "required": [
                "genres",
                "movie_type",
                "thumbnail_url",
                "title"
            ],
            "properties": {
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "demon",
                        "fantasy",
                        "supernatural"
                    ]
                },
                "movie_type": {
                    "type": "string",
                    "example": "anime"
                },
                "producer": {
                    "type": "string",
                    "example": "Agate"
                },
                "rating": {
                    "type": "string",
                    "example": "5.9"
                },
                "release_date": {
                    "type": "string",
                    "example": "2023-05-01"
                },
                "status": {
                    "type": "string",
                    "example": "complete"
                },
                "studio": {
                    "type": "string",
                    "example": "Agate"
                },
                "synopsis": {
                    "type": "string",
                    "example": "Agate is a demon girl cursed with an eternal life. To forget the past, she throws half of her head into a deep valley and runs away. Unexpectedly, the tears flood the valley floor and form into a lake, triggering a flood in hell. Aloys, the Prince of Ghost comes to the Human World to find out the truth and finally finds the demon girl, starting a story of life and love."
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://s4.anilist.co/file/anilistcdn/media/anime/cover/medium/bx141914-P1xQHMXN7q6z.png"
                },
                "title": {
                    "type": "string",
                    "example": "Wu Nao Monü"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model.MovieDetails"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithPaginate-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model.MovieDetails"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model.MovieDetails": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "movie_id": {
                    "type": "string"
                },
                "movie_type": {
                    "type": "string"
                },
                "producer": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "studio": {
                    "type": "string"
                },
                "synopsis": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "description": "Di title harus sama seperti source anilist atau sebgainya, karena title ini akan digunakan sebagai key finding",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}